# adsapi/static/openapi.yaml
openapi: 3.0.0
info:
  title: Test API
  description: A test API to show ads
  version: 0.0.1

servers:
  - url: https://macken.dafnet.se/david/
    description: External testserver
  - url: http://localhost:8081/
    description: Local testserver

paths:
  /ads:
    get:
      summary: Returns a list of all ads
      parameters:
        - in: query
          name: sort
          required: false
          schema:
            type: string
            enum:
              - price
              - created
            default: created
          description: The value to sort the ads by. Define order with `order`.
        - in: query
          name: order
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
          description: The order to sort the results
      responses:
        '200':
          description: A JSON array of ads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ad'
        '400':
          description: Error message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Inserts an Ad in the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ad'
      responses:
        '201':
          description: Ad added OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          description: Error message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /ads/{id}:
    get:
      summary: Returns a single ad by id
      parameters:
        - $ref: '#/components/parameters/AdId'
      responses:
        '200':
          description: A JSON object of an ad
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ad'
    delete:
      summary: Deletes an ad from the database
      parameters:
        - $ref: '#/components/parameters/AdId'
      responses:
        '201':
          $ref: '#/components/schemas/Success'

components:
  parameters:
    AdId:
      in: path
      name: id
      required: true
      description: The ad id
      schema:
        type: string
        format: uuid

  schemas:
    Ad:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        subject:
          type: string
        body:
          type: string
        price:
          type: number
          format: float
          minimum: 0
        email:
          type: string
          format: email
        created:
          type: string
          format: date-time
          readOnly: true
      required:
        - subject
        - body
        - email
    Error:
      type: object
      properties:
        code:
          type: integer
        name:
          type: string
        description:
          type: string
    Success:
      type: object
      properties:
        success:
          type: boolean